Properly Format &'s
I was having issues trying to scrape based on a tag with an & in it, and i noticed that & translates into %26 in the url. There was nothing replacing it in the code, so I added a line to do so and it now works. Opening a whole issue for something so minor seemed like a waste of your time, so I did this instead